hawks_batter_server/requirements.txt
flask==2.0.1
flask-smorest==0.34.1
flask-cors==3.0.10
PyYAML==6.0.1
RPi.GPIO==0.7.1
pynput==1.7.6
#pythohn-vlc==3.0.20123
# Dev requirements
#black ==22.3.0
#pylint ==2.14.3
